crawling-jobplanet/requirements.txt
requests
beautifulsoup4
selenium
lxml
python-dotenv
webdriver-manager
psycopg2-binary
rapidfuzz
psutil
